requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0
psycopg2-binary==2.9.7

# Forms and Validation
WTForms==3.1.1
email-validator==2.1.0

# File uploads (Flask-Reuploaded is the maintained fork)
Flask-Reuploaded==1.3.0

# Security
bcrypt==4.1.2
cryptography==41.0.7
python-dotenv==1.0.0

# Utilities
Pillow==10.1.0
unidecode==1.3.7
phonenumbers==8.13.29

# Development tools
autopep8==2.0.4
pycodestyle==2.11.1

# Payment integration
requests==2.31.0
pycryptodome==3.19.0

# Other
alembic==1.13.1
python-dateutil==2.8.2
MarkupSafe==2.1.3
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2

# Caching
Flask-Caching==2.1.0

# Image processing
Pillow==10.1.0

# System monitoring
psutil==5.9.6

# Optional - for search functionality
flask-msearch==0.2.9.4

# For PDF generation (if needed)
reportlab==4.0.7
